Change behavior of HeaderSection (#86)
HeaderSection was previously designed to display a header and list of items. So it used ItemSection and ListSection internally. Now HeaderSection is modified to use ItemSection and NestedSection internally. This allows us to add any type of section below the header. Necessary changes are made in test cases and sample apps as well.